grant and grant opportunity datasets

A dataset that contains the information about grants opportunity details

Usage

grant_opp

Format

A data frame with 2843 rows and 25 variables.

funding_opportunity_title

Title of the opportunity

expected_number_of_awards

Expected count of awards

current_closing_date_for_applications

When the opportunity is currently scheduled to close

award_ceiling

Maximum individual award amount in dollars

award_floor

Minimum individual award amount in dollars

category

category or type of field that the grant opportunity belongs to

opportunity_id

Integer ID for this opportunity

estimated_total_program_funding

Estimated funding amount in dollars

eligibility_individuals

Are individuals eligible?

eligibility_state_governments

Are state governments eligible?

eligibility_county_governments

Are county governments eligible?

eligibility_independent_school_districts

Are independent school districts eligible?

eligibility_city_or_township_governments

Are city or township governments eligible?

eligibility_special_district_governments

Are special district governments eligible?

eligibility_native_american_tribal_governments_federally_recognized

Are Native American tribal governments (Federally recognized) eligible?

eligibility_native_american_tribal_organizations_other

Are Native American tribal organizations (other than Federally recognized tribal governments) eligible?

eligibility_nonprofits_501c3

Are nonprofits having a 501(c)(3) status with the IRS, other than institutions of higher education eligible?

eligibility_nonprofits_non_501c3

Are nonprofits that do not have a 501(c)(3) status with the IRS, other than institutions of higher education eligible?

eligibility_for_profit

Are for profit organizations other than small businesses eligible?

eligibility_small_businesses

Are small businesses eligible?

eligibility_private_institutions_of_higher_education

Are private institutions of higher education eligible?

eligibility_public_institutions_of_higher_education

Are public and State controlled institutions of higher education eligible?

eligibility_public_indian_housing_authorities

Are public housing authorities and Indian housing authorities eligible?

eligibility_others

Are other groups eligible?

eligibility_unrestricted

Is eligibility unrestricted?

Source

https://github.com/rfordatascience/tidytuesday/tree/master/data/2023/2023-10-03